chore: fix race condition
