+++ apps/cowswap-frontend/src/modules/trade/hooks/useTradeState.ts
@@ -8,15 +8,16 @@ import { useLimitOrdersRawState, useUpdateLimitOrdersRawState } from 'modules/li
 import { useSwapRawState, useUpdateSwapRawState } from 'modules/swap/hooks/useSwapRawState'
 import { ExtendedTradeRawState, TradeRawState } from 'modules/trade/types/TradeRawState'
 
-import { useTradeTypeInfo } from './useTradeTypeInfo'
-
 import { TradeType } from '../types'
+import { useTradeTypeInfoFromUrl } from './useTradeTypeInfoFromUrl'
+
+const EMPTY_TRADE_STATE = {}
 
 export function useTradeState(): {
   state?: TradeRawState
   updateState?: (update: Partial<ExtendedTradeRawState>) => void
 } {
-  const tradeTypeInfo = useTradeTypeInfo()
+  const tradeTypeInfo = useTradeTypeInfoFromUrl()
 
   const limitOrdersState = useLimitOrdersRawState()
   const updateLimitOrdersState = useUpdateLimitOrdersRawState()
@@ -28,7 +29,7 @@ export function useTradeState(): {
   const updateSwapState = useUpdateSwapRawState()
 
   return useMemo(() => {
-    if (!tradeTypeInfo) return {}
+    if (!tradeTypeInfo) return EMPTY_TRADE_STATE
 
     if (tradeTypeInfo.tradeType === TradeType.SWAP) {
       return {
